sizes

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-	printf("size of a char: %lu bytes\n", sizeof(char));
-	printf("size of an int: %lu bytes\n", sizeof(int));
-	printf("size of a long int: %lu bytes\n", sizeof(long int));
-	printf("size of a long long int: %lu bytes\n", sizeof(long long int));
-	printf("size of a float: %lu bytes\n", sizeof(float));
+	printf("size of a char: %lu bytes\n", (unsigned long)sizeof(char));
+	printf("size of an int: %lu bytes\n", (unsigned long)sizeof(int));
+	printf("size of a long int: %lu bytes\n", (unsigned long)sizeof(long int));
+	printf("size of a long long int: %lu bytes\n", (unsigned long)sizeof(long long int));
+	printf("size of a float: %lu bytes\n", (unsigned long)sizeof(float));
 	return (0);
 }
